testing

+++ src/testing.py
@@ -45,7 +45,7 @@ def createsample(folderpath):
         imgt = torch.from_numpy(img).permute(2, 0, 1)
         coord = coords(H=720, W=1280)
         imgt = torch.cat((imgt, coord), dim=0) 
-        imgs.append((img, imgt.unsqueeze(dim=0)))
+        imgs.append((img, imgt.unsqueeze(dim=0).float()))
     
     return imgs
